refactor(www): group trust capability
